ULM-integrated version of KEVM (#19)
This modifies the KEVM repository so that the semantics can be plugged
directly into Geth. This version of the semantics is not yet passing all
tests of the ethereum test suite, nor can it be tested using KEVM's
integration test harness, so we are not merging it into `master`.
However, I would like it to be code reviewed, so I created this PR.