remove some nondeterminism in driver.md (#22)
There are a few rules in driver.md that conflict with each other or
which conflict with rules in state-utils.md. It is pretty
straightforward to modify the semantics so that this is no longer the
case. We do this in order to test that the EVM semantics is actually
100% deterministic (which it is).