Tweaks to evm semantics to make summarizaion easier.
We introduce the ==Bytes construct, because K seems to have difficulty
figuring out that a term of sort K passed reached via executing the semantics
is the same as that introduced via pyk. The reason for this seems related
to the first being parsed as a KSequence, i.e. giveing us `X:Bytes ~> .K`.
I am not sure which is "correct", given the dual nature of the K sort
as the top sort, vs as a sequence.

We also change the requires clause for KZGPOINTEVAL's unhappy path to make it obvious to the backend
that is the negation of that for the happy path.